+++ tailormap-components/projects/core/src/lib/user-interface/attributelist/attributelist-common/attributelist-statistic.spec.ts
@@ -1,0 +1,7 @@
+import { AttributelistStatistic } from './attributelist-statistic';
+
+describe('AttributelistStatistic', () => {
+  it('should create an instance', () => {
+    expect(new AttributelistStatistic()).toBeTruthy();
+  });
+});
